requirements.txt
allure-pytest==2.8.18
allure-python-commons==2.8.18
attrs==21.4.0
configparser==5.2.0
enum34==1.1.10
iniconfig==1.1.1
lxml==4.8.0
namedlist==1.8
nitratexml2dict==0.2.2
packaging==21.3
pluggy==1.0.0
py==1.11.0
pyparsing==3.0.7
pytest==7.1.1
pyxml2dict==1.0.3
PyYAML==6.0
six==1.16.0
tomli==2.0.1
XML2Dict==0.2.2
